refactor(desktop): move showBrowserViewContextMenu to tab-state (BP-P5b-6)
Pulls in 2 more deps (queueBrowserDownloadPrompt + browserContextSuggestedFilename)
and the electron Menu/clipboard imports. queue/consume DownloadOverride and
browserContextSuggestedFilename stay in main.ts as bind helpers — they touch
workspace.pendingDownloadOverrides and a path-resolver that haven't migrated.

main.ts: 22,252 → 22,116 (-136 lines)
